+++ src/scss/components/_header.scss
@@ -22,4 +22,16 @@
       font-size: 28px;
     }
   }
+
+  &__modal{
+    padding: 5px 20px;
+    background-color: $colorWhite;
+    border: solid $colorLightGray 1.5px;
+    border-radius: 10px;
+
+    &:hover{
+      background-color: $colorLightGray;
+      border: solid $colorWhite 2px;
+    }
+  }
 }
